Небольшой фикс конфига
