--- src/main/java/faang/school/postservice/config/redis/RedisSerializerConfig.java
@@ -1,23 +1,0 @@
-package faang.school.postservice.config.redis;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializer;
-
-@Configuration
-public class RedisSerializerConfig {
-
-    @Bean
-    public RedisSerializer<Object> jackson2JsonRedisSerializer() {
-        ObjectMapper objectMapper = new ObjectMapper()
-                .registerModule(new JavaTimeModule())
-                .setSerializationInclusion(JsonInclude.Include.NON_NULL)
-                .findAndRegisterModules();
-
-        return new Jackson2JsonRedisSerializer<>(objectMapper, Object.class);
-    }
-}
